move contents of restore_test.go into other files and delete it
